separating tab bar styles from components + importing styles from styles file
