+++ src/TabBar.js
@@ -20,6 +20,8 @@ import type {
   Style,
 } from './TabViewTypeDefinitions';
 
+import styles from './tabBarStyles.js'
+
 type IndicatorProps<T> = SceneRendererProps<T> & {
   width: Animated.Value,
 };
@@ -499,59 +501,3 @@ export default class TabBar<T: Route<*>> extends React.PureComponent<
     );
   }
 }
-
-const styles = StyleSheet.create({
-  container: {
-    flex: 1,
-  },
-  scroll: {
-    overflow: Platform.OS === 'web' ? 'auto' : 'scroll',
-  },
-  tabBar: {
-    backgroundColor: '#2196f3',
-    elevation: 4,
-    shadowColor: 'black',
-    shadowOpacity: 0.1,
-    shadowRadius: StyleSheet.hairlineWidth,
-    shadowOffset: {
-      height: StyleSheet.hairlineWidth,
-    },
-    // We don't need zIndex on Android, disable it since it's buggy
-    zIndex: Platform.OS === 'android' ? 0 : 1,
-  },
-  tabContent: {
-    flexDirection: 'row',
-    flexWrap: 'nowrap',
-  },
-  tabLabel: {
-    backgroundColor: 'transparent',
-    color: 'white',
-    margin: 8,
-  },
-  tabItem: {
-    flexGrow: 1,
-    padding: 8,
-    alignItems: 'center',
-    justifyContent: 'center',
-  },
-  badge: {
-    position: 'absolute',
-    top: 0,
-    right: 0,
-  },
-  indicatorContainer: {
-    position: 'absolute',
-    top: 0,
-    left: 0,
-    right: 0,
-    bottom: 0,
-  },
-  indicator: {
-    backgroundColor: '#ffeb3b',
-    position: 'absolute',
-    left: 0,
-    bottom: 0,
-    right: 0,
-    height: 2,
-  },
-});
